refactor: reorganize directories and improve code readability and separation of concerns
- Reorganized project directories for better structure.
- Enhanced code readability.
- Improved separation of responsibilities across the codebase.
