+++ package.json
@@ -5,8 +5,10 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-  "start": "nodemon index.js",
-  "test": "echo \"Error: no test specified\" && exit 1"
+    
+    "build": "tsc",
+    "start": "nodemon index.js",
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
   "author": "",
@@ -16,9 +18,7 @@
     "axios": "^1.7.5",
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
-    "express": "^4.19.2",
-    "llamaai": "^1.0.4",
-    "openai": "^4.56.0"
+    "express": "^4.19.2"
   },
   "devDependencies": {
     "@eslint/js": "^9.6.0",
@@ -27,8 +27,8 @@
     "eslint-config-standard": "^17.1.0",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-n": "^16.6.2",
-    "standard": "^17.1.0",
+    "eslint-plugin-promise": "^6.4.0",
     "globals": "^15.8.0",
-    "eslint-plugin-promise": "^6.4.0"
+    "standard": "^17.1.0"
   }
 }
