+++ src/constants/promptConstants.js
@@ -1,0 +1,9 @@
+const TRAINING_LEVELS = {
+  BEGGINER: 'Principiante',
+  ADVANCED: 'Avanzado',
+  MEDIUM: 'Intermedio'
+};
+
+export {
+  TRAINING_LEVELS
+};
